Develop

+++ controller/users.js
@@ -4,10 +4,17 @@ const User = require('../models/user');
 const {
   validateUser, isAValidEmail, isAWeakPassword,
 } = require('../utils/utils');
-// const { isAdmin } = require('../middleware/auth');
+const { isAdmin } = require('../middleware/auth');
+const { pagination } = require('../utils/utils');
 
 // GET '/users'
 const getUsers = (req, res, next) => {
+  const url = `${req.protocol}://${req.get('host') + req.path}`;
+  const options = {
+    page: parseInt(req.query.page, 10) || 1,
+    limit: parseInt(req.query.limit, 10) || 2,
+  };
+
   const userFind = User.find({});
   userFind
     .then((doc) => {
@@ -15,6 +22,9 @@ const getUsers = (req, res, next) => {
         return next(404);
       }
       if (doc) {
+        const links = pagination(doc, url, options.page, options.limit, doc.totalPages);
+        res.links(links);
+        console.log(doc);
         return res.status(200).send(doc);
       }
     })
@@ -25,12 +35,19 @@ const getUsers = (req, res, next) => {
 
 // GET 'users:id'
 const getOneUser = async (req, res, next) => {
-  const userid = req.params.uid;
-  await User.findById(userid, (err, userfound) => {
-    if (err) return next(404);
-    // if (!userfound) return res.status(404).send({ message: 'El usuario no ha sido encontrado' });
-    res.status(200).send(userfound);
-  });
+  try {
+    const { uid } = req.params;
+    const value = validateUser(uid);
+    const user = await User.findOne(value).lean();
+    if (!user) {
+      return next(404);
+    }
+
+    if (req.authToken.uid === user._id.toString() || isAdmin(req)) return res.json(user);
+    return next(403);
+  } catch (err) {
+    return next(err);
+  }
 };
 
 // POST '/users'
@@ -42,7 +59,7 @@ const newUser = async (req, res, next) => {
       return next(400);
     }
 
-    // if (isAWeakPassword(password) || !isAValidEmail(email)) return next(400);
+    if (isAWeakPassword(password) || !isAValidEmail(email)) return next(400);
 
     const findUser = await User.findOne({ email: req.body.email });
 
@@ -55,8 +72,8 @@ const newUser = async (req, res, next) => {
     const newUser = new User(req.body);
     newUser.password = bcrypt.hashSync(req.body.password, 10);
     const userSaved = await newUser.save(newUser);
-    // const user = await User.findOne({ _id: userSaved._id }).select('-password');
-    return res.status(200).json(userSaved);
+    const user = await User.findOne({ _id: userSaved._id }).select('-password');
+    return res.status(200).json(user);
   } catch (err) {
     next(err);
   }
@@ -75,8 +92,8 @@ const updateUser = async (req, res, next) => {
 
     if (!user) return next(404);
 
-    // if (req.authToken.uid !== user._id.toString() && !isAdmin(req)) return next(403);
-    // if (!isAdmin(req) && body.roles) return next(403);
+    if (req.authToken.uid !== user._id.toString() && !isAdmin(req)) return next(403);
+    if (!isAdmin(req) && body.roles) return next(403);
     if (Object.entries(body).length === 0) return next(400);
 
     if (body.password && isAWeakPassword(body.password)) return next(400);
@@ -107,7 +124,7 @@ const deleteOneUser = async (req, res, next) => {
 
     if (!userDeleted) return next(404);
 
-    // if (req.authToken.uid !== userDeleted._id.toString() && !isAdmin(req)) return next(403);
+    if (req.authToken.uid !== userDeleted._id.toString() && !isAdmin(req)) return next(403);
 
     await User.findOneAndDelete(value);
